mcps/jira/pyproject.toml
[project]
name = "jira-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.0.2",
    "python-dotenv>=1.0",
    "atlassian-python-api>=4.0",
]
